refactor the file structure
