+++ Salwador/UIComponents/GenerateImage/GenerateImageView.swift
@@ -12,7 +12,6 @@ struct GenerateImageView: View {
     @State private var prompText: String = ""
     @State private var isPopL: String = ""
     @StateObject var viewModel = GenerateImageViewModel()
-    @State var isPopListShown: Bool = false
     @State var generatedImage = UIImage()
 
     var body: some View {
@@ -22,12 +21,11 @@ struct GenerateImageView: View {
                 textEditorView
                 submitButtonView
                 Spacer()
-
                 imageView
                 Spacer()
             }
             .background(Color("BackgroundColor"))
-            if isPopListShown {
+            if viewModel.isPopListShown {
                 ZStack(alignment: .bottom) {
                     PopList
                 }
@@ -70,7 +68,7 @@ private extension GenerateImageView {
             viewModel.buttonText = "Wait..."
             viewModel.isLoading = true
             withAnimation {
-                isPopListShown = false
+                viewModel.isPopListShown = false
             }
 
             Task {
@@ -109,7 +107,7 @@ private extension GenerateImageView {
                             print("Long pressed!")
                             withAnimation {
                                 generatedImage = identifier
-                                isPopListShown = true
+                                viewModel.isPopListShown = true
                             }
                         }
                 }
@@ -120,97 +118,5 @@ private extension GenerateImageView {
                 }
             }
         }
-    }
-
-    var loadingView: some View {
-        VStack {
-            ProgressView()
-                .foregroundColor(.white)
-            Text("Your image is being generated...")
-                .font(.title3)
-                .foregroundColor(Color("TextColor"))
-        }
-    }
-
-    var PopList: some View {
-        ScrollView(.vertical, showsIndicators: false) {
-            VStack {
-                VStack {
-                    HStack {
-                        Button {
-                            withAnimation {
-                                UIImageWriteToSavedPhotosAlbum(generatedImage, nil, nil, nil)
-                                isPopListShown = false
-                            }
-                        }
-                    label: {
-                            Image(systemName: "square.and.arrow.down")
-                                .resizable()
-                                .frame(width: 22, height: 25)
-                                .foregroundColor(.white)
-                            Text("Save")
-                                .fontWeight(.bold)
-                                .foregroundColor(.white)
-                        }
-                        Spacer()
-                    }
-                    .padding(.leading)
-
-                    Divider()
-
-                    HStack {
-                        let photo = Photo(image: Image(uiImage: generatedImage))
-
-                        ShareLink(
-                            item: photo.image,
-                            preview: SharePreview(
-                                "Hey, check it out! I've created an AI-Image via Salwador App...",
-                                image: photo.image
-                            )
-                        ) {
-                            HStack {
-                                Image(systemName: "square.and.arrow.up")
-                                    .resizable().frame(width: 22, height: 25).foregroundColor(.white)
-
-                                Text("Share")
-                                    .fontWeight(.bold)
-                                    .foregroundColor(.white)
-                            }
-                        }
-                        .labelStyle(.titleAndIcon)
-                        .imageScale(.large)
-                        .symbolVariant(.fill)
-                        .foregroundColor(.white)
-                        Spacer()
-                    }
-                    .padding(.leading)
-                }
-                .padding(10)
-                .background(Color("OrangeColor")
-                )
-                .cornerRadius(10)
-
-                ZStack {
-                    Button {
-                        withAnimation {
-                            isPopListShown = false
-                        }
-                    } label: {
-                        HStack {
-                            Image(systemName: "xmark")
-                                .resizable()
-                                .frame(width: 15, height: 15)
-                                .foregroundColor(Color("OrangeColor")
-                                )
-                                .padding(10)
-                        }
-                    }
-                }
-                .background(.white)
-                .clipShape(Circle())
-            }
-        }
-        .padding(.top)
-        .frame(width: 200, height: 190)
     }
 }
